code/visualize/requirements.txt
asn1crypto==0.24.0
chumpy==0.70
click==7.1.2
cryptography==2.3
dataclasses==0.7
decorator==4.4.2
fastapi==0.61.1
ffmpeg-python==0.2.0
freetype-py==2.2.0
future==0.18.2
h11==0.11.0
idna==2.6
imageio==2.9.0
keyring==10.6.0
keyrings.alt==3.0
-e /workspace/code
networkx==2.5
numpy==1.19.2
Pillow==8.0.1
pycrypto==2.6.1
pydantic==1.6.1
pyglet==1.5.8
pygobject==3.26.1
git+http://github.com/mmatl/pyopengl.git
pyrender==0.1.43
pyxdg==0.26
scipy==1.5.3
SecretStorage==2.3.1
six==1.11.0
smplx==0.1.26
starlette==0.13.6
toml==0.10.1
torch==1.4.0
torchgeometry==0.1.2
tqdm==4.50.2
trimesh==3.8.12
typing-extensions==3.7.4.3
uvicorn==0.12.2
